fix: fixed urls and refactored secrets.ts file
- changed abalobi.live and abalobi.info to abalobi.org
- The fact that secrets.ts is a local copy (because it is excluded in
.gitignore) for developers means there are different versions floating
around. The secrets.js file only contains urls which aren't necessarily
secret. Therefore the secrets.ts has been renamed to base_urls.ts which
contain the correct urls used when this app is deployed. This file is
not in .gitignore which means all clones of the repository should have
the same urls.